Day 6 / Part 2 - Solution.

+++ AdventOfCode-2021/src/UnitTests/Day6Tests.fs
@@ -395,10 +395,10 @@ module UnitTests.Day6Tests
 
         let result = Some lanternFishSeq |> Part2.toLanternFishDictionary
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(1uy, 1)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(1uy, 1UL)
             
-        Assert.Equal<Dictionary<byte, int>>(
+        Assert.Equal<Dictionary<byte, uint64>>(
             expected,
             result)
 
@@ -412,10 +412,10 @@ module UnitTests.Day6Tests
 
         let result = Some lanternFishSeq |> Part2.toLanternFishDictionary
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(1uy, 2)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(1uy, 2UL)
             
-        Assert.Equal<Dictionary<byte, int>>(
+        Assert.Equal<Dictionary<byte, uint64>>(
             expected,
             result)
 
@@ -430,179 +430,179 @@ module UnitTests.Day6Tests
 
         let result = Some lanternFishSeq |> Part2.toLanternFishDictionary
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(1uy, 2)
-        expected.Add(2uy, 1)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(1uy, 2UL)
+        expected.Add(2uy, 1UL)
             
-        Assert.Equal<Dictionary<byte, int>>(
+        Assert.Equal<Dictionary<byte, uint64>>(
             expected,
             result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when empty dictionary then returns empty dictionary`` () =
-        let lanternFish = Dictionary<byte, int>()
+        let lanternFish = Dictionary<byte, uint64>()
 
         let result = lanternFish |> Part2.spawnLanternFish
 
-        Assert.Equal<Dictionary<byte, int>>(Dictionary<byte, int>(), result)
+        Assert.Equal<Dictionary<byte, uint64>>(Dictionary<byte, uint64>(), result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when single lanternFish with DaysUntilSpawn = 8 then returns a single lanternFish with DaysUntilSpan = 7`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(8uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(8uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFish
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(7uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(7uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when single lanternFish with DaysUntilSpawn = 7 then returns a single lanternFish with DaysUntilSpan = 6`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(7uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(7uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFish
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(6uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(6uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when single lanternFish with DaysUntilSpawn = 6 then returns a single lanternFish with DaysUntilSpan = 5`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(6uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(6uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFish
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(5uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(5uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when single lanternFish with DaysUntilSpawn = 5 then returns a single lanternFish with DaysUntilSpan = 4`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(5uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(5uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFish
         
-        let expected = Dictionary<byte, int>()
-        expected.Add(4uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(4uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when single lanternFish with DaysUntilSpawn = 4 then returns a single lanternFish with DaysUntilSpan = 3`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(4uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(4uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFish
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(3uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(3uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when single lanternFish with DaysUntilSpawn = 3 then returns a single lanternFish with DaysUntilSpan = 2`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(3uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(3uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFish
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(2uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(2uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when single lanternFish with DaysUntilSpawn = 2 then returns a single lanternFish with DaysUntilSpan = 1`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(2uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(2uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFish
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(1uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(1uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when single lanternFish with DaysUntilSpawn = 1 then returns a single lanternFish with DaysUntilSpan = 0`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(1uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(1uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFish
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(0uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(0uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when single lanternFish with DaysUntilSpawn = 0 then returns two lanternFish one with DaysUntilSpan = 6 and one with DaysUntilSpawn = 6`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(0uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(0uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFish
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(6uy, 1)
-        expected.Add(8uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(6uy, 1UL)
+        expected.Add(8uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFish: when initial state is initial state from problem 3,4,3,1,2 then result is 2,3,2,0,1`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(3uy, 2)
-        lanternFish.Add(4uy, 1)
-        lanternFish.Add(1uy, 1)
-        lanternFish.Add(2uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(3uy, 2UL)
+        lanternFish.Add(4uy, 1UL)
+        lanternFish.Add(1uy, 1UL)
+        lanternFish.Add(2uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFish
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(2uy, 2)
-        expected.Add(3uy, 1)
-        expected.Add(0uy, 1)
-        expected.Add(1uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(2uy, 2UL)
+        expected.Add(3uy, 1UL)
+        expected.Add(0uy, 1UL)
+        expected.Add(1uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFishTimes: when initial state is initial state lanternFish is empty dictionary then result is empty dictionary`` () =
-        let lanternFish = Dictionary<byte, int>()
+        let lanternFish = Dictionary<byte, uint64>()
 
         let result = lanternFish |> Part2.spawnLanternFishTimes <| 1
 
-        Assert.Equal<Dictionary<byte, int>>(Dictionary<byte, int>(), result)
+        Assert.Equal<Dictionary<byte, uint64>>(Dictionary<byte, uint64>(), result)
 
     [<Fact>]
     let ``Part2.spawnLanternFishTimes: when initial state is initial state lanternFish 0 and times is 1 then result is 8,6`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(0uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(0uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFishTimes <| 1
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(6uy, 1)
-        expected.Add(8uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(6uy, 1UL)
+        expected.Add(8uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFishTimes: when initial state is initial state lanternFish 0 and times is 2 then result is 7,5`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(0uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(0uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFishTimes <| 2
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(5uy, 1)
-        expected.Add(7uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(5uy, 1UL)
+        expected.Add(7uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
 
     [<Fact>]
     let ``Part2.spawnLanternFishTimes: when initial state is initial state lanternFish 0 and times is 3 then result is 6,4`` () =
-        let lanternFish = Dictionary<byte, int>()
-        lanternFish.Add(0uy, 1)
+        let lanternFish = Dictionary<byte, uint64>()
+        lanternFish.Add(0uy, 1UL)
 
         let result = lanternFish |> Part2.spawnLanternFishTimes <| 3
 
-        let expected = Dictionary<byte, int>()
-        expected.Add(4uy, 1)
-        expected.Add(6uy, 1)
-        Assert.Equal<Dictionary<byte, int>>(expected, result)
+        let expected = Dictionary<byte, uint64>()
+        expected.Add(4uy, 1UL)
+        expected.Add(6uy, 1UL)
+        Assert.Equal<Dictionary<byte, uint64>>(expected, result)
